chore: refactor + coherence
-refactored folder structure for the Sensors project
- made sure the iot.proto in shared is the same as the one in grpc-server
