+++ internet_of_things/Sensors/src/main.c
@@ -3,9 +3,9 @@
 #include <stdio.h>
 #include "uart_stdio.h"
 
-#include "sensorRead.h" // access interface
+#include "services/sensorRead.h" // access interface
 #include "light.h"
-#include "communication.h"
+#include "services/communication.h"
 
 #include "wifi.h" // Include WiFi driver
 #define USE_WIFI_COMM 0 // Change this to 1 when ready to use WiFi
