+++ internet_of_things/Sensors/src/services/communication.h
