+++ shared/IoT.proto
@@ -5,15 +5,17 @@ syntax = "proto3";
 package iotService;
 
 enum sensorType {
-    temp = 0;
-    light = 1;
-    water = 2;
-    error = 3; 
+    error = 0;
+    temp = 1;
+    light = 2;
+    water = 3;
+    fillCup = 4;
 }
 service iotService {
     rpc getTemperature(tempReq) returns (tempRes);
     rpc getLight(lightReq) returns (lightRes);
     rpc getWaterLevel(waterLevelReq) returns (waterLevelRes);
+    rpc fillCup(fillCupReq) returns (fillCupRes);
 
 }
 
@@ -43,6 +45,14 @@ message waterLevelRes {
     Status status = 2;
 }
 
+message fillCupReq {
+    int32 arduinoId = 1;
+    optional float amount = 2; //amount of water to fill in milliliters, optional because we can also just request a standard quantity fill, set up previously
+}
+message fillCupRes {
+     Status status = 1;
+}
+
 message Status {
     bool success = 1;
     string message = 2;
@@ -53,3 +63,7 @@ message sensorReading {
     sensorType type = 2; //to identify the kind of sensor. possibly use an enum
     int64 timestamp = 3;
 }
+
+
+
+
